move types out of state to dedicated module
