+++ app/packages/playback/index.ts
@@ -1,5 +1,6 @@
 export * from "./src/lib/constants";
 export * from "./src/lib/hooks";
 export * from "./src/lib/state";
+export * from "./src/lib/types";
 export * from "./src/lib/utils";
 export * from "./src/views/Timeline";
